pyproject.toml
[project]
name = "rsi-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "yfinance>=0.2.66",
]
